+++ frontend/src/component/segments/CreateSegment/CreateSegment.tsx
@@ -10,9 +10,9 @@ import useToast from 'hooks/useToast';
 import { useNavigate } from 'react-router-dom';
 import { formatUnknownError } from 'utils/formatUnknownError';
 import { useSegmentForm } from '../hooks/useSegmentForm';
-import { SegmentForm } from '../SegmentForm/SegmentForm';
+import { SegmentForm } from '../SegmentForm';
 import { feedbackCESContext } from 'component/feedback/FeedbackCESContext/FeedbackCESContext';
-import { segmentsDocsLink } from 'component/segments/SegmentDocs/SegmentDocs';
+import { segmentsDocsLink } from 'component/segments/SegmentDocs';
 import { useSegmentValuesCount } from 'component/segments/hooks/useSegmentValuesCount';
 import { SEGMENT_CREATE_BTN_ID } from 'utils/testIds';
 import { useSegmentLimits } from 'hooks/api/getters/useSegmentLimits/useSegmentLimits';
