+++ frontend/src/component/segments/EditSegment/EditSegment.tsx
@@ -11,10 +11,10 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { formatUnknownError } from 'utils/formatUnknownError';
 import { useSegmentForm } from '../hooks/useSegmentForm';
-import { SegmentForm } from '../SegmentForm/SegmentForm';
+import { SegmentForm } from '../SegmentForm';
 import { segmentsFormDescription } from 'component/segments/CreateSegment/CreateSegment';
 import { UpdateButton } from 'component/common/UpdateButton/UpdateButton';
-import { segmentsDocsLink } from 'component/segments/SegmentDocs/SegmentDocs';
+import { segmentsDocsLink } from 'component/segments/SegmentDocs';
 import { useSegmentValuesCount } from 'component/segments/hooks/useSegmentValuesCount';
 import { SEGMENT_SAVE_BTN_ID } from 'utils/testIds';
 import { useSegmentLimits } from 'hooks/api/getters/useSegmentLimits/useSegmentLimits';
